The new System
work on the index.php file. I'm trying to ftp it now for testing
purposes haha but the schools network keeps timing out so i'll have to
do it from home. I know there is alot of files here. There is the
index.php, login.php, register.php, addFood.php and search.php that are
the ones we created.  the rest are from the template so dont worry about
them lol those are for design purposes.